[FIX] Padronizando nomenclatura
