--- l10n_br_contabilidade/views/account_mapeamento.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2018 ABGF
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-
-<openerp>
-    <data>
-
-        <record id="l10n_br_account_mapeamento_form_view" model="ir.ui.view">
-            <field name="name">l10n_br.conta.mapeamento.form</field>
-            <field name="model">account.mapeamento</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <h2><field name="name" nolabel="1"/></h2>
-                        <group>
-                            <field name="referencia_plano_id" attrs="{'required': True}" options="{'no_quick_create': True, 'no_create_edit': True}"/>
-                            <field name="conta_sistema_id" attrs="{'required': True}"/>
-                            <field name="conta_plano_conta_externo" attrs="{'required': True}"/>
-                            <field name="mapeamento_pai_id"/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-
-        <record id="l10n_br_account_mapeamento_de_para_tree_view" model="ir.ui.view">
-            <field name="name">l10n_br.conta.mapeamento.tree</field>
-            <field name="model">account.mapeamento</field>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="referencia_plano_id"/>
-                    <field name="conta_sistema_id"/>
-                    <field name="conta_plano_conta_externo"/>
-                    <field name="mapeamento_pai_id"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="l10n_br_account_mapeamento_form" model="ir.actions.act_window">
-            <field name="name">Mapeamento de Contas</field>
-            <field name="res_model">account.mapeamento</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <record id="l10n_br_account_mapeamento_referencia_form_view" model="ir.ui.view">
-            <field name="name">l10n_br.conta.mapeamento.referencia.form</field>
-            <field name="model">account.mapeamento.referencia</field>
-            <field name="arch" type="xml">
-                <form>
-                    <sheet>
-                        <group>
-                            <field name="name" attrs="{'required': True}"/>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-
-        <record id="l10n_br_account_mapeamento_referencia_tree_view" model="ir.ui.view">
-            <field name="name">l10n_br.conta.mapeamento.referencia.tree</field>
-            <field name="model">account.mapeamento.referencia</field>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="name"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="l10n_br_contabilidade_account_mapeamento_referencia_action" model="ir.actions.act_window">
-            <field name="name">Referência Plano de Contas</field>
-            <field name="res_model">account.mapeamento.referencia</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-    </data>
-</openerp>
